fix: import WorldGDP section Framer Motion's components on client-side
