+++ src/components/DataRoom/WorldGDP/SlidingVideo.tsx
@@ -1,0 +1,10 @@
+import MotionTypography from '@/components/common/MotionTypography'
+import css from './styles.module.css'
+
+const SlidingVideo = () => (
+  <MotionTypography animateYFrom={-30} customDelay={0.2}>
+    <video className={css.video} src="/videos/DataRoom/SafeGlobe.mp4" autoPlay muted loop />
+  </MotionTypography>
+)
+
+export default SlidingVideo
