+++ src/components/DataRoom/WorldGDP/styles.module.css
@@ -13,17 +13,6 @@
   padding: 108px 36px 0;
 }
 
-.video {
-  width: 400px;
-  opacity: 0.8;
-}
-
-.text {
-  font-size: 16px;
-  line-height: 24px;
-  align-self: flex-start;
-}
-
 .contentContainer {
   display: flex;
   flex-direction: column;
@@ -31,10 +20,21 @@
   align-items: center;
 }
 
-.percentContainer {
+.video {
+  width: 400px;
+  opacity: 0.8;
+}
+
+.text {
+  width: 240px;
+  font-size: 16px;
+  line-height: 24px;
+  align-self: flex-start;
+  margin-left: 36px;
+}
+
+.textContainer {
   margin-top: 24px;
-  display: flex;
-  gap: 12px;
 }
 
 .linksContainer {
@@ -44,28 +44,26 @@
 }
 
 @media (max-width: 900px) {
+  .stickyContainer {
+    padding: 72px 30px 0;
+    align-items: center;
+  }
+
   .contentContainer {
     flex-grow: 1;
   }
 
-  .stickyContainer {
-    padding: 72px 30px 0;
-    align-items: center;
+  .video {
+    width: 300px;
+  }
+
+  .textContainer {
+    margin-top: 12px;
   }
 
   .linksContainer {
     position: static;
     margin: 0;
     align-self: flex-start;
-  }
-
-  .percentContainer {
-    margin-top: 12px;
-    flex-direction: column;
-    gap: 0px;
-  }
-
-  .video {
-    width: 300px;
   }
 }
